Refactor ArtistsContainer
- Move ArtistsContainer to ../containers/Artists
