+++ src/containers/Artists/Artists.js
@@ -1,9 +1,9 @@
 import React, {Component} from "react";
 
-class ArtistsContainer extends Component {
+class Artists extends Component {
   render() {
     return <div>ArtistsContainer</div>;
   }
 }
 
-export default ArtistsContainer;
+export default Artists;
